Move betty.plugins.css_resource to betty.css_resources
